tradingview-webhook/requirements.txt
requests
binance-futures-connector